+++ .babelrc
@@ -1,0 +1,16 @@
+{
+    "presets": ["@babel/preset-env"],
+    "plugins": [
+      [
+        "@babel/plugin-transform-runtime",
+        {
+          "absoluteRuntime": false,
+          "corejs": false,
+          "helpers": true,
+          "regenerator": true,
+          "useESModules": true
+        }
+      ],
+      ["@babel/plugin-proposal-export-default-from"]
+    ]
+  }
