+++ .npmignore
@@ -1,0 +1,4 @@
+src
+samples
+yarn.lock
+.nvmrc
