+++ package.json
@@ -1,16 +1,17 @@
 {
   "name": "cluster-scheduler",
-  "version": "0.0.1",
+  "version": "0.0.2",
   "description": "Holds an in-memory queue for scheduling tasks to be delegated to cluster workers",
-  "main": "src/js/index.js",
+  "engines": {
+    "node": ">=6.0.0"
+  },
+  "main": "dist/js/index.js",
   "repository": "https://github.com/christherama/cluster-scheduler.git",
   "author": "Chris Ramacciotti",
   "license": "MIT",
   "scripts": {
-    "build": "babel src --presets=@babel/preset-env --out-dir=dist --source-maps",
-    "start": "node ./dist/js/index.js"
+    "build": "rm -rf dist && babel src -d dist"
   },
-  "type": "module",
   "dependencies": {
     "@babel/polyfill": "^7.7.0",
     "winston": "^3.2.1"
@@ -18,6 +19,9 @@
   "devDependencies": {
     "@babel/cli": "^7.7.0",
     "@babel/core": "^7.7.0",
-    "@babel/preset-env": "^7.7.1"
+    "@babel/plugin-proposal-export-default-from": "^7.5.2",
+    "@babel/plugin-transform-runtime": "^7.6.2",
+    "@babel/preset-env": "^7.7.1",
+    "create-index": "^2.5.0"
   }
 }
