+++ src/js/queue.js
@@ -16,4 +16,4 @@ class Queue {
   }
 }
 
-export default Queue;
+export { Queue };
